+++ services/monitor/src/Monitor.ts
@@ -1,7 +1,10 @@
 import DecentralizedStorageFetcher from "./DecentralizedStorageFetcher";
 import assert from "assert";
 import { EventEmitter } from "stream";
-import type { FetchRequestRPC } from "@ethereum-sourcify/lib-sourcify";
+import type {
+  FetchRequestRPC,
+  SourcifyRpc,
+} from "@ethereum-sourcify/lib-sourcify";
 import { SourcifyChain } from "@ethereum-sourcify/lib-sourcify";
 import logger from "./logger";
 import "./loggerServer"; // Start the dynamic log level server
@@ -39,9 +42,17 @@ export default class Monitor extends EventEmitter {
 
     this.config = deepMerge(defaultConfig, passedConfig || {});
 
+    const monitorFactoriesEnv = process.env["MONITOR_FACTORIES"];
+    if (monitorFactoriesEnv !== undefined) {
+      const monitorFactories =
+        monitorFactoriesEnv.toLowerCase() === "true" ? true : false;
+      this.config.monitorFactories = monitorFactories;
+    }
     logger.info(
-      "Starting the monitor using the effective config: " +
-        JSON.stringify(this.config, null, 2), // Stringify here to see the full config clearly
+      `Monitoring factories is ${this.config.monitorFactories ? "enabled" : "disabled"}`,
+      {
+        monitorFactories: this.config.monitorFactories,
+      },
     );
 
     if (this.config.decentralizedStorages?.ipfs?.enabled) {
@@ -69,9 +80,7 @@ export default class Monitor extends EventEmitter {
       } else {
         return new SourcifyChain({
           chainId: chain.chainId,
-          rpcs: authenticateRpcs(chain).map((rpc) => ({
-            rpc,
-          })),
+          rpcs: authenticateRpcs(chain),
           name: chain.name,
           supported: true,
         });
@@ -101,7 +110,10 @@ export default class Monitor extends EventEmitter {
       }
     }
 
-    if (chainsInConfigButNotChainsToMonitor.length > 0) {
+    if (
+      chainsInConfigButNotChainsToMonitor.length > 0 &&
+      process.env.NODE_ENV === "production"
+    ) {
       throw new Error(
         `Chain configs found for chains that are not being monitored: ${chainsInConfigButNotChainsToMonitor.join(
           ",",
@@ -145,9 +157,7 @@ export default class Monitor extends EventEmitter {
   };
 }
 
-export function authenticateRpcs(
-  chain: MonitorChain,
-): (string | FetchRequestRPC)[] {
+export function authenticateRpcs(chain: MonitorChain): SourcifyRpc[] {
   return chain.rpc.map((rpc) => {
     if (typeof rpc === "string") {
       if (rpc?.includes("ethpandaops.io")) {
@@ -165,9 +175,9 @@ export function authenticateRpcs(
             },
           ],
         };
-        return fetchRequestRpc;
+        return { rpc: fetchRequestRpc };
       }
-      return rpc;
+      return { rpc };
     }
     if (rpc?.type === "ApiKey") {
       const apiKey = process.env[rpc.apiKeyEnvName] || "";
@@ -176,7 +186,21 @@ export function authenticateRpcs(
           `API key ${rpc.apiKeyEnvName} not found in environment variables`,
         );
       }
-      return rpc.url.replace("{API_KEY}", apiKey);
+      let secretUrl = rpc.url.replace("{API_KEY}", apiKey);
+      if (rpc?.subDomainEnvName) {
+        // subDomain is optional
+        const subDomain = process.env[rpc.subDomainEnvName];
+        if (!subDomain) {
+          throw new Error(
+            `Subdomain ${rpc.subDomainEnvName} not found in environment variables`,
+          );
+        }
+        secretUrl = secretUrl.replace("{SUBDOMAIN}", subDomain);
+      }
+      return {
+        rpc: secretUrl,
+        ...(rpc.traceSupport ? { traceSupport: rpc.traceSupport } : {}),
+      };
     }
     throw new Error("Invalid rpc object: " + JSON.stringify(rpc));
   });