+++ services/server/test/chains/chain-tests.spec.ts
@@ -2015,6 +2015,30 @@ describe("Test Supported Chains", function () {
     "perfect",
   );
 
+  // Jovay Sepolia Testnet
+  verifyContract(
+    "0x63A2fA72305de66654bB95606190f5477abeAf36",
+    "2019775",
+    "Jovay Sepolia Testnet",
+    "shared/",
+  );
+  // Jovay Mainnet
+  verifyContract(
+    "0x63A2fA72305de66654bB95606190f5477abeAf36",
+    "5734951",
+    "Jovay Mainnet",
+    "shared/",
+  );
+
+  // Tempo Testnet (Moderato)
+  verifyContract(
+    "0xba5Ed099633D3B313e4D5F7bdc1305d3c28ba5Ed",
+    "42431",
+    "Tempo Testnet (Moderato)",
+    "createx/",
+    "partial",
+  );
+
   it("should have included Etherscan contracts for all testedChains having etherscanAPI", function (done) {
     const missingEtherscanTests: ChainApiResponse[] = [];
     supportedChains